% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-gam_mod.R
\name{gam_mod}
\alias{gam_mod}
\title{General Interface for GAM Models}
\usage{
gam_mod(mode = "regression")
}
\arguments{
\item{mode}{A single character string for the type of model.}
}
\description{
General Interface for GAM Models
}
