% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-gam_mod.R
\name{gen_additive_mod_fit_impl}
\alias{gen_additive_mod_fit_impl}
\title{Low-Level GAM function for translating modeltime to forecast}
\usage{
gen_additive_mod_fit_impl(
  formula,
  data,
  k_term1 = NULL,
  k_term2 = NULL,
  k_term3 = NULL,
  k_term4 = NULL,
  k_term5 = NULL,
  ...
)
}
\arguments{
\item{formula}{A dataframe of xreg (exogenous regressors)}

\item{data}{A numeric vector of values to fit}

\item{...}{Additional arguments passed to \code{forecast::Arima}}
}
\description{
Low-Level GAM function for translating modeltime to forecast
}
