% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-gen_additive_mod_params.R
\name{gen_additive_mod}
\alias{gen_additive_mod}
\alias{smoother_dim_term1}
\alias{smoother_dim_term2}
\alias{smoother_dim_term3}
\alias{smoother_dim_term4}
\alias{smoother_dim_term5}
\title{Tuning Parameters for GAMS Models}
\usage{
smoother_dim_term1(range = c(0L, 10L), trans = NULL)

smoother_dim_term2(range = c(0L, 10L), trans = NULL)

smoother_dim_term3(range = c(0L, 10L), trans = NULL)

smoother_dim_term4(range = c(0L, 10L), trans = NULL)

smoother_dim_term5(range = c(0L, 10L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Tuning Parameters for GAMS Models
}
\details{
The main parameters for ARIMA models are:
\itemize{
\item \code{smoother_dim}: The the dimension of the basis used to represent a smooth term.
}
}
\examples{
smoother_dim()


}
