% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-gam_mod.R
\name{gen_additive_mod_predict_impl}
\alias{gen_additive_mod_predict_impl}
\title{Bridge prediction Function for GAMS Models}
\usage{
gen_additive_mod_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Additional arguments passed to \code{stats::predict()}}
}
\value{
A vector of values (predictions) with class \code{numeric}.
}
\description{
Bridge prediction Function for GAMS Models
}
